requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
python-dotenv==1.0.0
cryptography==41.0.7
pyjwt==2.8.0
qrcode==7.4.2
Werkzeug==2.3.7
waitress==2.1.2
python-jose==3.3.0
PyNaCl==1.5.0
base58==2.1.1